// libs/foundation/dashboard/src/lib/shared/models/request-filter.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText } from '@gosi-ui/core';
export class RequestFilter {
  status: BilingualText[] = [];
  type: BilingualText[] = [];
  channel: BilingualText[] = [];
  legalEntity: BilingualText[] = [];
  filedOffice: BilingualText[] = [];
  village: BilingualText[] = [];
  villageId: number[] = [];
}